data_source_ingestion/app/requirements.txt
pandas
sqlalchemy
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
psycopg2-binary
cryptography